# Improved perfomance of builtin objects, initial changes that is done for discussion