As was discussed in this issue:

https://github.com/brython-dev/brython/issues/1419

`Brython` is less efficinet than JavaScript ...
It is due to Object.$B.getattrib called more times

Native types like Array in V8 is optimized

`Python` list object is not inherited from other object that mean we could optimize default behavior of object with getting attribute and instead append and remove from native `JavaScript Array` object

Before optimization:
```console
javascript: 2.59716796875ms
brython.js:5316 using indexedDB for stdlib modules cache
brython.js:9073 python: 289.85302734375ms
```

After optimization:
```console
javascript: 2.6259765625ms
brython.js:5243 using indexedDB for stdlib modules cache
brython.js:9031 python: 51.942138671875ms
```

I would like to discuss this optimization after initial commit ... to not uselessly spent my time )