@schmave 2020-06-17
Cool! I imagine that most of the speedup here is coming from not calling `$B.args`--is that true? $B.args() generates an appropriate error message when this function is called with the wrong number of arguments, so that functionality will have to be replaced somehow. Maybe you could optimize $B.args, or create a alternative to it for simple functions like this that don't take any keyword args.

I don't know what the effect of __proto__ is. Does that speed things up too?

I don't normally chime in on PRs, but you happened to have pick two of my favorite topics (speed optimizations and working inside Node). I'm sure Pierre will have thoughts too.