@schmave 2020-06-17
Also in case you aren't aware of them, the project does have a fairly large set of automated tests that you can run to verify that this change doesn't break something unexpected: https://github.com/brython-dev/brython/blob/master/CONTRIBUTING.md#testing