@redradist 2020-06-17
Also seems like assignment does not decrease performance ...
I commented it out and nothing happens with performance