@redradist 2020-06-18
> Significant improvement is to work with JavaScript Array-s directly and add simple mapping:
> 
>     1. Python list -> JavaScript Array
> 
>     2. Python set -> JavaScript Object
> 
>     3. Python dict -> JavaScript Object
>        ...
>        N) Custom Python object that inherits from some other class -> As it is now without modifications
> 
> 
> In such way we could improve performance with Native to the Browser Types, but still work with Python ;)

Actually I will argue with myself )))
We should not use JavaScript Object instead Python, it would be less Pythonic and `cause the issue is not in Native objects it would not make sense

What would make sense is to get simple positional argument first without calling `$B.args` method each time, because it slow and not needed for positional argument

Also significant improvements could be done by caching `$B.$getattr` (already did some work) ...
It could be proven by doing such thing in `Python` code:
```
            from browser import console, window

            console.time('python')
            lst = list()
            i = 0
            append = lst.append
            while i < 100000:
                append(i)
                i += 1
            console.timeEnd('python')
```

After such small modifications perfomance improves even more by 50%:
```console
javascript: 2.258056640625ms
brython.js:5243 using indexedDB for stdlib modules cache
brython.js:9054 python: 26.261962890625ms
```

And now from 300ms -> 26ms ... much better, but we need to apply the same tech in `$B.$getattr`, not to search for `attr` each time, but take a cached value if `attr` was not changed