@PierreQuentel 2020-06-18
Thanks for the PR. I really wish I could merge it beacause it dramatically improves the speed of `list.append()` in the most usual case, but it is not compliant with Python in other cases, as shown in the following examples.

CPython

```
Python 3.8.0 (tags/v3.8.0:fa919fd, Oct 14 2019, 19:37:50) [MSC v.1916 64 bit (AMD64)] on win32
Type "help", "copyright", "credits" or "license" for more information.
>>> t = []
>>> t.append(1)
>>> t
[1]
>>> t.append(2, 3)
Traceback (most recent call last):
  File "<stdin>", line 1, in <module>
TypeError: append() takes exactly one argument (2 given)
>>> t.append(4, x=5)
Traceback (most recent call last):
  File "<stdin>", line 1, in <module>
TypeError: append() takes no keyword arguments
>>> t.append(6)
>>> t.append(x=6)
Traceback (most recent call last):
  File "<stdin>", line 1, in <module>
TypeError: append() takes no keyword arguments
>>> t
[1, 6]
>>>
```

Brython console with the changes you proposed (I ran __`make_dist.py`__ to generate the correct version of __`brython.js`__)

```
Brython 3.8.9 on Netscape 5.0 (Windows)
>>> t = []
>>> t.append(1)
>>> t
[1]
>>> t.append(2, 3) # should raise TypeError
>>> t
[1, 2]
>>> t.append(4, x=5)  # should raise TypeError
>>> t
[1, 2, 4]
>>> t.append(x=6)  # should raise TypeError
>>> t
Traceback (most recent call last):
Exception: <Javascript TypeError>: obj is undefined
>>>
```