@redradist 2020-06-18
I have also improved performance of `$B.$getattr` method https://github.com/brython-dev/brython/pull/1425

With applying all this optimization `Brython` works only 10x times slower than JavaScript

I have tested it on the following example:
```html
        <script>
            console.time('javascript')
            var list = [];
            for (var i = 0; i <= 10000000; i++) {
                list.push(i);
            }
            console.timeEnd('javascript')
        </script>
        <script type="text/python">
            from browser import console, window

            console.time('python')
            lst = list()
            i = 0
            while i < 10000000:
                lst.append(i)
                i += 1
            console.timeEnd('python')
        </script>
```